CategorySpecificFilters uses useLabelSearch, type alignment
