+++ src/components/_experiment/searchFilters/SearchFilters.tsx
@@ -3,7 +3,7 @@ import { ChevronDown, ChevronRight, ChevronUp, Circle, ListFilter, SlidersHorizo
 import { useMemo, useState } from "react";
 
 import { Checkbox } from "@/components/atoms/checkbox/Checkbox";
-import { TLabelResult } from "@/hooks/useLabelSearch";
+import { TSearchLabel } from "@/types";
 import { hasPublishedDateRule } from "@/utils/_experiment/dateRangeFilters";
 import { labelTypeLabel } from "@/utils/_experiment/labelTypeLabel";
 import { joinTailwindClasses } from "@/utils/tailwind";
@@ -16,7 +16,7 @@ function hasValue(group: TQueryGroup | null | undefined, value: string): boolean
   return group.filters.some((f) => ("value" in f ? f.value === value : hasValue(f, value)));
 }
 
-function hasActiveFilterOfType(filters: TLabelResult[], group: TQueryGroup | null | undefined, type: TLabelType): boolean {
+function hasActiveFilterOfType(filters: TSearchLabel[], group: TQueryGroup | null | undefined, type: TLabelType): boolean {
   if (!group) return false;
   return group.filters.some((f) =>
     "value" in f ? filters.some((label) => label.id === f.value && label.type === type) : hasActiveFilterOfType(filters, f, type)
@@ -28,7 +28,7 @@ const OTHER_LABEL_TYPES = ["activity_status", "agent"] as const;
 export type TLabelType = (typeof PRIMARY_LABEL_TYPES)[number] | (typeof OTHER_LABEL_TYPES)[number];
 
 type TProps = {
-  availableFilters: TLabelResult[];
+  availableFilters: TSearchLabel[];
   filters?: TQueryGroup | null;
   activeLabelType: TLabelType;
   onActiveLabelTypeChange: (labelType: TLabelType) => void;
@@ -59,7 +59,7 @@ export function SearchFilters({
     [availableFilters, activeLabelType]
   );
 
-  const renderCheckboxRow = (filter: TLabelResult, isAvailable: boolean) => {
+  const renderCheckboxRow = (filter: TSearchLabel, isAvailable: boolean) => {
     const checked = hasValue(filters, filter.id);
     return (
       <li key={filter.id}>
